+++ src/components/composables.ts
@@ -3,6 +3,8 @@ import {
   signInWithEmailAndPassword,
   createUserWithEmailAndPassword,
 } from "firebase/auth";
+import { collection, getDocs } from "firebase/firestore";
+import { db } from "../main";
 
 const auth = getAuth();
 
@@ -34,6 +36,37 @@ export const register = (email: any, password: any) =>
     .catch((error) => {
       switch (error.code) {
         case "auth/invalid-email":
-         return "Invalid email";
+          return "Invalid email";
       }
     });
+
+export const getFeedPaints = () =>
+  getDocs(collection(db, "users")).then((docs) => {
+    let resDocs: any = [];
+    docs.forEach((doc) => {
+      doc.data().paints.forEach((paint: any) => {
+        resDocs.push({
+          nameOfPaint: paint.nameOfPaint,
+          date: new Date(paint.date.seconds * 1000).toLocaleDateString(),
+          userName: doc.data().name,
+        });
+      });
+    });
+    resDocs = resDocs
+      .sort(
+        (prev: any, curr: any) => Date.parse(prev.date) - Date.parse(curr.date)
+      )
+      .reverse();
+    return resDocs;
+  });
+
+export const getUserPaints = () => {
+  let paintsResult: any = [];
+  getDocs(collection(db, "users")).then((docs) => {
+    docs.forEach((doc) => {
+      if (doc.data().name === localStorage.getItem("email"))
+        paintsResult = doc.data().paints;
+    });  
+  });
+  return paintsResult;
+};
